moving the pause intent out of the download info
